Avoid hidden virtual function warnings
The base class declaration of value() is const.  This isn't a bug
because we're not actually using value() in these derived classes but
the compiler can't be sure of that.  (what I can't be sure of, on the
other hand, is why these classes derive from FunctionBase in the first
place if they're not going to use its interface...)