Mention optional external PETSc/libMesh builds
I don't think using an external libMesh build would be a good idea for
most users, but lots of systems already have a PETSc built, and on those
systems using the existing PETSc might be convenient (e.g. if it took
special configuration for a system-specific MPI install, or if you just
want to get up and running faster) or unavoidable (if it got installed
to a library path that gets picked up by the libMesh or MOOSE build
systems via other dependencies).